Revert making pipeline driver a class.
This change reduces the conceptual complexity of ap_verify, and
makes the data flow more obvious. Dependencies on the ap_pipe API
are still contained in a separate module, where they can't clutter
up the top-level logic.
